+++ .storybook/main.js
@@ -1,3 +1,6 @@
+const codesandbox = require('remark-codesandbox');
+const { resolve } = require('path');
+const palmettoComponentsCodesandboxTemplatePackageJSON = require('../docs/codesandbox-template/package.json');
 /**
  * The majority of the development webpack config comes directly from the @storybook
  * default config, since the output is rendered by their platform.
@@ -5,7 +8,6 @@
  * and if so, generate additional config options that will get merged into the storybook config
  * in order to process Typescript components and SCSS files appropriately.
  */
-
 const webpackConfig = require('../webpack.config');
 
 module.exports = {
@@ -37,6 +39,36 @@ module.exports = {
     './register',
   ],
   webpackFinal: config => {
+    const mdxRule = config.module.rules.find((rule) =>
+      rule.test.test('.story.mdx')
+    );
+
+    const {
+      options: { remarkPlugins }
+    } = mdxRule.use.find(
+      ({ loader }) => loader === require.resolve('@mdx-js/loader')
+    );
+
+    remarkPlugins.push([
+      codesandbox,
+      {
+        mode: 'iframe',
+        query: {
+          fontsize: 14
+        },
+        customTemplates: {
+          'palmetto-components': {
+            extends: `file:${resolve(
+              __dirname,
+              '../docs/codesandbox-template'
+            )}`,
+            entry: 'src/MyComponent.js'
+          },
+        },
+        autoDeploy: true
+      }
+    ]);
+
     return {
       ...config,
       module: {