+++ src/components/Button/Button.tsx
@@ -2,7 +2,8 @@ import React, {
   FC, ReactNode, MouseEvent, FocusEvent, forwardRef, createElement, AnchorHTMLAttributes,
 } from 'react';
 import classNames from 'classnames';
-import { IconName } from '../../types';
+import { IconName, ResponsiveProp } from '../../types';
+import { generateResponsiveClasses } from '../../lib/generateResponsiveClasses';
 import { handleReactRouterClick } from '../../lib/reactRouterClickHandler';
 import { Box } from '../Box/Box';
 import { Icon } from '../Icon/Icon';
@@ -96,7 +97,7 @@ export interface ButtonProps {
   /**
    * The size of the button.
    */
-  size?: ButtonSize;
+  size?: ButtonSize | ResponsiveProp<ButtonSize>;
   /**
    * The color variant of the button
    */
@@ -137,14 +138,22 @@ export const Button: FC<ButtonProps> = forwardRef(
   ) => {
     const disabled = isLoading || isDisabled;
 
-    const buttonClasses = classNames(styles.button, className, {
-      [styles.outline]: isOutlined && !isNaked,
-      [styles.loading]: isLoading,
-      [styles.naked]: isNaked,
-      [styles[variant]]: variant && !isNaked,
-      [styles[size]]: size && !isNaked,
-      [styles['full-width']]: fullWidth,
-    });
+    const responsiveClasses = generateResponsiveClasses('size', size).map(c => styles[c]);
+
+    const buttonClasses = classNames(
+      'palmetto-components__variables__button',
+      'palmetto-components__variables__form-control',
+      styles.button,
+      className,
+      ...!isNaked ? responsiveClasses : [],
+      {
+        [styles.outline]: isOutlined && !isNaked,
+        [styles.loading]: isLoading,
+        [styles.naked]: isNaked,
+        [styles[variant]]: variant && !isNaked,
+        [styles['full-width']]: fullWidth,
+      },
+    );
 
     const handleClick = handleReactRouterClick;