+++ src/components/FileUpload/FileUpload.tsx
@@ -96,7 +96,7 @@ export interface FileUploadProps extends BoxProps {
   /**
    * Color for button component. Matches a curated subset of button variants.
    */
-  variant?: 'light' | 'dark';
+  variant?: 'light' | 'dark' | 'primary';
   /**
    * Additional props to be spread. IMPORTANT: these will be spread ONLY to the
    * `input` element in the component since it is the actual semantic file input
@@ -133,7 +133,8 @@ export const FileUpload: FC<FileUploadProps> = React.forwardRef<HTMLDivElement, 
 ) => {
   const hiddenFileInput = useRef<HTMLInputElement>(null);
 
-  const handleClick = () => {
+  const handleClick = (e: MouseEvent) => {
+    e.stopPropagation();
     if (hiddenFileInput?.current) hiddenFileInput.current.click();
   };
 
@@ -196,15 +197,25 @@ export const FileUpload: FC<FileUploadProps> = React.forwardRef<HTMLDivElement, 
           size={size}
           fullWidth={fullWidth}
         >
-          {hasIcon && (
-            <Icon
-              name="upload"
-              className="m-right-xs align-self-center"
-              data-testid="file-upload__upload-icon"
-            />
-          )}
-          {buttonText}
-          {isRequired && <>&nbsp;*</>}
+          {/* We need the onClick handler here to prevent bubbling of clicks on the inner button elements */}
+          {/* Implemented in response to Safari's handling of bubbled click events into the details element */}
+          {/* which triggered default behavior of opening the details element on click */}
+          <Box as="span" direction="row" onClick={(e: MouseEvent) => { e.preventDefault(); }}>
+            {hasIcon && (
+              <Icon
+                name="upload"
+                className={classNames('align-self-center', { 'm-right-xs': buttonText })}
+                data-testid="file-upload__upload-icon"
+              />
+            )}
+            {buttonText
+              && (
+                <Box data-testid="file-upload__upload-text">
+                  {buttonText}
+                </Box>
+              )}
+            {isRequired && <>&nbsp;*</>}
+          </Box>
           <Box
             // We spread props here at that top to avoid inputProps overwriting high-level component props
             {...inputProps}