+++ src/components/Tabs/TabItem.tsx
@@ -7,33 +7,33 @@ export interface TabItemProps extends BoxProps {
 
 export const TabItem: React.FC<TabItemProps> = ({
   children,
+  as = 'button',
   isDisabled = false,
   padding = 'md',
-  style,
+  style = {},
   ...restProps
 }) => {
   const styles = {
     ...style,
     cursor: !isDisabled ? 'pointer' : 'not-allowed',
+    background: style?.background || 'transparent',
+    whiteSpace: style?.whiteSpace || 'nowrap',
   };
 
   return (
-    <Box style={{ ...styles }} role="tab" {...restProps}>
-      <Box
-        as="button"
-        disabled={isDisabled}
-        style={{
-          background: 'transparent',
-          whiteSpace: 'nowrap',
-        }}
-        padding={padding}
-        borderWidth="0"
-        color={isDisabled ? 'grey-100' : 'grey-400'}
-        direction="row"
-        justifyContent="center"
-      >
-        {children}
-      </Box>
+    <Box
+      style={{ ...styles }}
+      role="tab"
+      as={as}
+      disabled={isDisabled}
+      padding={padding}
+      borderWidth="0"
+      color={isDisabled ? 'grey-100' : 'grey-400'}
+      direction="row"
+      justifyContent="center"
+      {...restProps}
+    >
+      {children}
     </Box>
   );
 };